Rename helpers package to config
