+++ cmd/arc/services/k8s_watcher.go
@@ -4,7 +4,7 @@ import (
 	"fmt"
 	"log/slog"
 
-	cfg "github.com/bitcoin-sv/arc/internal/helpers"
+	cfg "github.com/bitcoin-sv/arc/internal/config"
 	"github.com/bitcoin-sv/arc/internal/k8s_watcher"
 	"github.com/bitcoin-sv/arc/internal/k8s_watcher/k8s_client"
 	"github.com/bitcoin-sv/arc/pkg/blocktx"
